+++ test/mocks/user_mock.js
@@ -1,7 +1,7 @@
 import { RoleManagerMock } from "./role_manager_mock.js";
-import {Permissions} from 'discord.js';
-const DEFAULT_PERMS = [Permissions.FLAGS.SEND_MESSAGES, Permissions.FLAGS.VIEW_CHANNEL, Permissions.FLAGS.READ_MESSAGE_HISTORY, 
-Permissions.FLAGS.CHANGE_NICKNAME, Permissions.FLAGS.ADD_REACTIONS];
+import { Permissions } from 'discord.js';
+const DEFAULT_PERMS = [ Permissions.FLAGS.SEND_MESSAGES, Permissions.FLAGS.VIEW_CHANNEL, Permissions.FLAGS.READ_MESSAGE_HISTORY, 
+    Permissions.FLAGS.CHANGE_NICKNAME, Permissions.FLAGS.ADD_REACTIONS ];
 
 class UserMock {
     
@@ -10,19 +10,19 @@ class UserMock {
         this.roles = new RoleManagerMock(guild);
         if (data) {
             for (const property in data) {
-                this.property = data[property];
+                this[property] = data[property];
             }
         }
-        this.permissions = new Permissions(DEFAULT_PERMS)
+        this.permissions = new Permissions(DEFAULT_PERMS);
         
     }
 
 
     hasPermission(tags, options) {
-        if(options.checkAdmin && this.permissions.has(Permissions.FLAGS.ADMINISTRATOR)) {
+        if (options.checkAdmin && this.permissions.has(Permissions.FLAGS.ADMINISTRATOR)) {
             return true;
         }
-        return this.permissions.has(tags)
+        return this.permissions.has(tags);
     }
 }